Trait luajit::types::LuaValue

pub trait LuaValue {
    fn push_val(self, l: *mut lua_State);
}

Represents any value that can be pushed onto the Lua stack

Required Methods

fn push_val(self, l: *mut lua_State)

push_val should push the value of this type the the top of the stack on Lua state l.